requirements.txt
## Core
pyyaml>=6.0
click>=8.0
rich>=13.0
tqdm>=4.65

# Video download
yt-dlp>=2024.1.0

# Subtitle download (dedicated, more reliable than yt-dlp subs)
youtube-transcript-api>=0.6.0    # YouTube subtitles
bilibili-api-python>=16.0.0     # Bilibili subtitles
aiohttp>=3.9.0                   # For bilibili-api async HTTP

# Audio processing
ffmpeg-python>=0.2.0
pydub>=0.25.1
soundfile>=0.12.0
numpy>=1.24.0

# Speech-to-text
faster-whisper>=1.0.0

# Translation backends
openai>=1.0.0           # For OpenAI/vLLM/SGLang (all OpenAI-compatible)
anthropic>=0.40.0       # For Claude API
requests>=2.31.0        # For Ollama
transformers>=4.36.0    # For HuggingFace local models
sentencepiece>=0.1.99   # For HuggingFace tokenizers

# TTS
edge-tts>=6.1.0

# Utilities
langcodes>=3.3.0
